refactor(#7): split track and callbacks DSLs for SCP
